+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Version 1.1 xx
 
-* Group layers according to legend interface in combo boxes
+* Plugin renamed to Raster Interpolation
+* Can choose band
 * Fix new raster API
+* Group layers according to legend interface in combo boxes
 * code respecting PEP8
 * use new versions of setting manager and combo manager
 
 ## Version 1.0 26.03.2013
+
+* Initial release of Land It plugin